refactor: children component
