+++ packages/gatsby-theme-api-docs/src/components/type/type.js
@@ -1,10 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FullWidthContainer } from '@commercetools-docs/gatsby-theme-docs';
+import {
+  FullWidthContainer,
+  SideBySide,
+} from '@commercetools-docs/gatsby-theme-docs';
+import SpacingsStack from '@commercetools-uikit/spacings-stack';
 import { generateTypeURN } from '../../utils/ctp-urn';
 import { useApiTypes } from '../../hooks/use-api-types';
 import reportError from '../../utils/report-error';
-import Children from './children';
+import Description from '../description';
+import Enum from './enum';
+import Properties from './properties/properties';
+import Examples from './examples';
 
 const ApiType = (props) => {
   const apiTypes = useApiTypes();
@@ -21,6 +28,26 @@ const ApiType = (props) => {
     );
   }
 
+  const DescriptionAndEnums = (props) => {
+    return (
+      <>
+        {props.apiType.description && (
+          <Description>{props.apiType.description}</Description>
+        )}
+        {props.apiType.enumeration && (
+          <Enum
+            values={props.apiType.enumeration}
+            enumDescriptions={props.apiType.enumDescriptions}
+          />
+        )}
+      </>
+    );
+  };
+
+  DescriptionAndEnums.propTypes = {
+    apiType: PropTypes.object.isRequired,
+  };
+
   const urn = generateTypeURN(matchedApiType);
 
   return (
@@ -28,16 +55,32 @@ const ApiType = (props) => {
       id={urn}
       aria-label={`${matchedApiType.displayName} definition`}
     >
-      <Children
-        apiKey={props.apiKey}
-        apiType={matchedApiType}
-        propertiesTableTitle={props.propertiesTableTitle}
-        renderDescriptionBelowProperties={
-          props.renderDescriptionBelowProperties
-        }
-        doNotRenderExamples={props.doNotRenderExamples}
-        hideInheritedProperties={props.hideInheritedProperties}
-      />
+      <SpacingsStack scale="m">
+        {!props.renderDescriptionBelowProperties && (
+          <DescriptionAndEnums apiType={matchedApiType} />
+        )}
+
+        {(matchedApiType.properties || matchedApiType.examples) && (
+          <SideBySide>
+            {matchedApiType.properties && (
+              <Properties
+                apiKey={props.apiKey}
+                apiType={matchedApiType}
+                title={props.propertiesTableTitle}
+                hideInheritedProperties={props.hideInheritedProperties}
+              />
+            )}
+
+            {matchedApiType.examples && !props.doNotRenderExamples && (
+              <Examples examples={matchedApiType.examples} />
+            )}
+          </SideBySide>
+        )}
+
+        {props.renderDescriptionBelowProperties && (
+          <DescriptionAndEnums apiType={matchedApiType} />
+        )}
+      </SpacingsStack>
     </FullWidthContainer>
   );
 };
